Add list package with concurrency-safe doubly linked list implementation.

+++ list/list_test.go
@@ -12,7 +12,7 @@ func TestNewList(t *testing.T) {
 	assert.NotNil(t, l)
 	assert.Nil(t, l.head)
 	assert.Nil(t, l.tail)
-	assert.Equal(t, uint(0), l.Len())
+	assert.Equal(t, 0, l.Len())
 }
 
 func TestPush(t *testing.T) {
@@ -21,11 +21,11 @@ func TestPush(t *testing.T) {
 	assert.NotNil(t, l.head)
 	assert.NotNil(t, l.tail)
 	assert.Equal(t, l.head, l.tail)
-	assert.Equal(t, uint(1), l.Len())
+	assert.Equal(t, 1, l.Len())
 	assert.Equal(t, 10, l.head.Element())
 
 	l.PushFront(20)
-	assert.Equal(t, uint(2), l.Len())
+	assert.Equal(t, 2, l.Len())
 	assert.Equal(t, 20, l.head.Element())
 	assert.Equal(t, 10, l.tail.Element())
 }
@@ -36,11 +36,11 @@ func TestAppend(t *testing.T) {
 	assert.NotNil(t, l.head)
 	assert.NotNil(t, l.tail)
 	assert.Equal(t, l.head, l.tail)
-	assert.Equal(t, uint(1), l.Len())
+	assert.Equal(t, 1, l.Len())
 	assert.Equal(t, 10, l.head.Element())
 
 	l.PushBack(20)
-	assert.Equal(t, uint(2), l.Len())
+	assert.Equal(t, 2, l.Len())
 	assert.Equal(t, 10, l.head.Element())
 	assert.Equal(t, 20, l.tail.Element())
 }
@@ -48,36 +48,36 @@ func TestAppend(t *testing.T) {
 func TestPopFront(t *testing.T) {
 	l := NewList[int]()
 	l.PopFront() // No-op
-	assert.Equal(t, uint(0), l.Len())
+	assert.Equal(t, (0), l.Len())
 
 	l.PushFront(10)
 	l.PopFront()
-	assert.Equal(t, uint(0), l.Len())
+	assert.Equal(t, (0), l.Len())
 	assert.Nil(t, l.head)
 	assert.Nil(t, l.tail)
 
 	l.PushFront(10)
 	l.PushFront(20)
 	l.PopFront()
-	assert.Equal(t, uint(1), l.Len())
+	assert.Equal(t, (1), l.Len())
 	assert.Equal(t, 10, l.head.Element())
 }
 
 func TestPopBack(t *testing.T) {
 	l := NewList[int]()
 	l.PopBack() // No-op
-	assert.Equal(t, uint(0), l.Len())
+	assert.Equal(t, (0), l.Len())
 
 	l.PushFront(10)
 	l.PopBack()
-	assert.Equal(t, uint(0), l.Len())
+	assert.Equal(t, (0), l.Len())
 	assert.Nil(t, l.head)
 	assert.Nil(t, l.tail)
 
 	l.PushFront(10)
 	l.PushFront(20)
 	l.PopBack()
-	assert.Equal(t, uint(1), l.Len())
+	assert.Equal(t, (1), l.Len())
 	assert.Equal(t, 20, l.head.Element())
 }
 
@@ -101,13 +101,13 @@ func TestBack(t *testing.T) {
 
 func TestLen(t *testing.T) {
 	l := NewList[int]()
-	assert.Equal(t, uint(0), l.Len())
+	assert.Equal(t, (0), l.Len())
 
 	l.PushFront(10)
-	assert.Equal(t, uint(1), l.Len())
+	assert.Equal(t, (1), l.Len())
 
 	l.PushBack(20)
-	assert.Equal(t, uint(2), l.Len())
+	assert.Equal(t, (2), l.Len())
 }
 
 func TestInsertAtPosition(t *testing.T) {
@@ -117,12 +117,12 @@ func TestInsertAtPosition(t *testing.T) {
 
 	err = l.InsertAtPosition(10, 0)
 	require.NoError(t, err)
-	assert.Equal(t, uint(1), l.Len())
+	assert.Equal(t, (1), l.Len())
 	assert.Equal(t, 10, l.head.Element())
 
 	err = l.InsertAtPosition(20, 1)
 	require.NoError(t, err)
-	assert.Equal(t, uint(2), l.Len())
+	assert.Equal(t, (2), l.Len())
 	assert.Equal(t, 20, l.tail.Element())
 
 	err = l.InsertAtPosition(30, 5)
@@ -130,7 +130,7 @@ func TestInsertAtPosition(t *testing.T) {
 
 	err = l.InsertAtPosition(25, 1)
 	require.NoError(t, err)
-	assert.Equal(t, uint(3), l.Len())
+	assert.Equal(t, (3), l.Len())
 	assert.Equal(t, 25, l.head.Next().Element())
 }
 
@@ -189,14 +189,14 @@ func TestIterateForward(t *testing.T) {
 func TestIterateBackward(t *testing.T) {
 	list := NewList[int]()
 	results := []struct {
-		index   uint
+		index   int
 		element int
 	}{}
 
 	// Case 1: Empty list
-	list.IterateBackward(func(index uint, element int) {
+	list.IterateBackward(func(index int, element int) {
 		results = append(results, struct {
-			index   uint
+			index   int
 			element int
 		}{index, element})
 	})
@@ -204,32 +204,32 @@ func TestIterateBackward(t *testing.T) {
 
 	// Case 2: Single element list
 	list.PushBack(42)
-	list.IterateBackward(func(index uint, element int) {
+	list.IterateBackward(func(index int, element int) {
 		results = append(results, struct {
-			index   uint
+			index   int
 			element int
 		}{index, element})
 	})
 	assert.Len(t, results, 1)
-	assert.Equal(t, uint(0), results[0].index)
+	assert.Equal(t, (0), results[0].index)
 	assert.Equal(t, 42, results[0].element)
 
 	// Case 3: Multiple elements
 	list.PushBack(100)
 	list.PushBack(200)
 	results = []struct {
-		index   uint
+		index   int
 		element int
 	}{} // Reset results
-	list.IterateBackward(func(index uint, element int) {
+	list.IterateBackward(func(index int, element int) {
 		results = append(results, struct {
-			index   uint
+			index   int
 			element int
 		}{index, element})
 	})
 	assert.Len(t, results, 3)
 	assert.Equal(t, []struct {
-		index   uint
+		index   int
 		element int
 	}{
 		{2, 200},
